requirements.txt
alembic==1.0.2
aniso8601==4.0.1
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.2
certifi==2018.10.15
chardet==3.0.4
Click==7.0
codeclimate-test-reporter==0.2.3
coverage==4.5.1
Flask==1.0.2
Flask-Migrate==2.3.0
flask-restplus==0.12.1
Flask-SQLAlchemy==2.3.2
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.16.3
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.8.0
psycopg2-binary==2.7.5
py==1.7.0
pycodestyle==2.4.0
pylint==2.1.1
pytest==3.10.0
pytest-cov==2.6.0
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.7
requests==2.20.1
six==1.11.0
SQLAlchemy==1.2.13
typed-ast==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
